requirements.txt
opencv-python
torch
torchvision
torchaudio
tensorboard
numpy
pandas
matplotlib
Pillow
SciencePlots
